// ngx-utils/src/lib/types/ngx-value.ts
export type NgxUtilsValue =
    | string
    | { type: 'BANK-CARD'; value: string; en?: true; join?: string }
    | { type: 'DATE'; value: Date; format?: string | 'FULL' | 'DATE' | 'TIME'; timezone?: string }
    | {
          type: 'DURATION';
          value: number | Date | { from: Date; to?: Date };
          en?: boolean;
          view?: 'FULL' | 'DAY' | 'HOUR';
      }
    | { type: 'EN'; value: string }
    | { type: 'MOBILE'; value: string; en?: true; join?: string }
    | { type: 'MULTILINE'; value: string; en?: boolean; html?: boolean }
    | { type: 'NUMBER'; value: number; en?: boolean };